% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneWayANCOVA.R
\docType{package}
\name{OneWayANCOVA}
\alias{OneWayANCOVA}
\title{Find test statistic value from a given data and corresponding critical value}
\description{
It gives test statistic and critical value for testing homogeneity of treatment effects against ordered alternatives in one-way ANCOVA model.
It also gives one sided confidence intervals for successive differences of treatment effects.
}
\author{
Anjana Mondal \email{mondal.anjana25@gmail.com}
}
